Drop fping and qstat from Recommends to Suggests
As discussed with Jan on IRC. fping and qstat are in universe in Ubuntu,
and requires an extensive delta to add an -extras package in universe
for Ubuntu users to benefit. Instead we can drop the Recommends to a
Suggests in Debian to save Ubuntu from having to do this.